backend/requirements.txt
-c constraints.txt
-e src/plonestudies[test]

zope.testrunner

# Add required add-ons
# (Ideally add them in setup.py for plonestudies)
# collective.easyform
